# backend/src/main/java/com/booki/domain/QuizAttempt.java
package com.booki.domain;

import jakarta.persistence.*;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import org.hibernate.annotations.CreationTimestamp;

import java.time.Instant;

@Entity
@Table(name = "quiz_attempts")
@Getter
@Setter
@NoArgsConstructor
public class QuizAttempt {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @ManyToOne(fetch = FetchType.LAZY)
    @JoinColumn(nullable = false, name = "session_id")
    private Session session;

    @Column(nullable = false, name = "page_number")
    private Integer pageNumber;

    @Column(nullable = false, columnDefinition = "LONGTEXT")
    private String question;

    @Column(nullable = false, columnDefinition = "LONGTEXT")
    private String answer;

    @Column(nullable = false, length = 20)
    private String difficulty;

    @ManyToOne(fetch = FetchType.LAZY)
    @JoinColumn(name = "profile_master_id")
    private ProfileMaster profileMaster;

    @Column(nullable = false)
    private Boolean correct;

    @Column(nullable = false)
    private Double score;

    @Column(nullable = false, columnDefinition = "LONGTEXT")
    private String feedback;

    @CreationTimestamp
    @Column(nullable = false, updatable = false, name = "created_at")
    private Instant createdAt;
}