# docs/openapi.yaml
openapi: 3.0.3
info:
  title: kobi-backend-api
  description: >
    API contract for BooKI's backend. This spec is reverse-engineered from the
    mock-backend's real, verified behavior (Express, in-memory) and from the
    frontend's existing types (`frontend/src/types/index.ts`) and endpoint map
    (`frontend/src/config/endpoints.ts`). It is meant to be the single source
    of truth while the real Spring Boot backend (`backend/`) is implemented,
    so both sides can be checked against the same shapes before writing more
    Java code.

    Endpoints that currently only exist in mock form (chat replies, quiz
    grading, summaries, PDF reports) are marked in their description as
    "mock logic" — the real backend is expected to replace the *logic*
    behind them (e.g. call a real AI provider) while keeping the *shape*
    of the request/response the same, so the frontend does not need to change.
  version: "1.0.0"
  license:
    name: Proprietary — internal use only
servers:
  - url: http://localhost:3001/api
    description: Mock backend (Node/Express, in-memory data)
  - url: http://localhost:8080/api
    description: Real backend (Spring Boot)
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Create an account
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                name: { type: string, description: "Defaults to the local part of the email if omitted" }
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/login:
    post:
      tags: [Auth]
      summary: Sign in
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /documents:
    get:
      tags: [Documents]
      summary: List my documents
      operationId: listDocuments
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Document" }
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Documents]
      summary: Upload a PDF
      operationId: uploadDocument
      description: >
        Real backend: extracts text per page (e.g. Apache PDFBox) and persists
        one `DocumentPage` row per page. Mock backend: ignores the uploaded
        file's real content and generates a few placeholder pages instead.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
        "400":
          $ref: "#/components/responses/BadRequest"
  /documents/{id}:
    get:
      tags: [Documents]
      summary: Get document metadata
      operationId: getDocument
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Documents]
      summary: Delete a document
      operationId: deleteDocument
      description: >
        Cascades: removes the document's pages, drops it from every Tag's
        membership list, and deletes any sessions (with their messages and
        quiz attempts) that pointed at it.
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /documents/{id}/file:
    get:
      tags: [Documents]
      summary: Download/stream the PDF file
      operationId: downloadDocumentFile
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: The PDF binary
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "404":
          $ref: "#/components/responses/NotFound"
...
  /collections/{id}/documents/{documentId}:
    put:
      tags: [Tags]
      summary: Add a document to a tag
      operationId: addDocumentToTag
      parameters:
        - $ref: "#/components/parameters/TagId"
        - $ref: "#/components/parameters/DocumentIdPath"
      responses:
        "200":
          description: Updated tag
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tag" }
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Tags]
      summary: Remove a document from a tag
      operationId: removeDocumentFromTag
      parameters:
        - $ref: "#/components/parameters/TagId"
        - $ref: "#/components/parameters/DocumentIdPath"
      responses:
        "200":
          description: Updated tag
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tag" }
        "404":
          $ref: "#/components/responses/NotFound"
  /profile-masters:
    get:
      tags: [Profile Masters]
      summary: List available Profile Masters
      operationId: listProfileMasters
      description: >
        Global, not per-user, in the current mock. A future version may scope
        custom Masters to their creator while keeping built-in ones shared.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProfileMaster" }
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Profile Masters]
      summary: Create a custom Profile Master
      operationId: createProfileMaster
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, systemPrompt]
              properties:
                name: { type: string }
                description: { type: string }
                systemPrompt: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProfileMaster" }
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /sessions/{id}:
    get:
      tags: [Sessions]
      summary: Get a session
      operationId: getSession
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "404":
          $ref: "#/components/responses/NotFound"
...
  /sessions/{id}/quiz:
    post:
      tags: [Quiz]
      summary: Generate quiz questions for this session
      operationId: generateQuiz
      description: >
        Mock logic: one templated question per page (up to `questionCount`),
        in the session's language. Real backend: could ask a real AI provider
        to generate genuine comprehension questions from the page text instead.
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                profileMasterId: { type: integer, nullable: true, description: "Defaults to the session's own Master" }
                difficulty: { $ref: "#/components/schemas/Difficulty" }
                questionCount: { type: integer, minimum: 1, maximum: 10, default: 3 }
      responses:
        "200":
          description: Generated quiz
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizGenerateResult" }
        "404":
          $ref: "#/components/responses/NotFound"
  /sessions/{id}/quiz/answer:
    post:
      tags: [Quiz]
      summary: Submit and grade one quiz answer
      operationId: submitQuizAnswer
      description: >
        Mock logic: keyword-overlap scoring against the page's text, with a
        pass threshold that gets stricter as `difficulty` increases. Every
        call is persisted as a `QuizAttempt`, which is what
        `GET /sessions/{id}/quiz/attempts` and the PDF quiz report read from.
        Real backend: replace the grading with a real AI judgment, keeping
        the `{correct, score, feedback}` shape.
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pageNumber, answer]
              properties:
                pageNumber: { type: integer }
                question: { type: string, description: "Echoed back verbatim into the stored QuizAttempt for the report" }
                answer: { type: string }
                difficulty: { $ref: "#/components/schemas/Difficulty" }
                profileMasterId: { type: integer, nullable: true }
      responses:
        "200":
          description: Grading result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizAnswerResult" }
        "404":
          $ref: "#/components/responses/NotFound"
  /sessions/{id}/quiz/attempts:
    get:
      tags: [Quiz]
      summary: Full quiz correction report for this session
      operationId: getQuizReport
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizReport" }
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        `Authorization: Bearer <token>`. The mock backend issues an unsigned
        placeholder token (base64 JSON, `alg: none`) — the real backend must
        issue a properly signed JWT instead, but the header format and the
        401-on-missing/invalid-or-expired-token behavior must stay the same.
  parameters:
    SessionId:
      name: id
      in: path
      required: true
      schema: { type: integer }
    DocumentId:
      name: id
      in: path
      required: true
      schema: { type: integer }
    DocumentIdPath:
      name: documentId
      in: path
      required: true
      schema: { type: integer }
    TagId:
      name: id
      in: path
      required: true
      schema: { type: integer }
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, invalid, or expired token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found, or not owned by the caller
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
    Difficulty:
      type: string
      enum: [easy, medium, hard]
    SessionLanguage:
      type: string
      enum: [en, es, fr]
      description: Language BooKI itself replies in for this session — independent of the language of the uploaded document.
...
    User:
      type: object
      properties:
        id: { type: integer }
        email: { type: string, format: email }
        name: { type: string }
        bio: { type: string }
        systemPrompt:
          type: string
          description: The "user context" layer — the reader's own stated learning preferences, applied across every session they create.
        createdAt: { type: string, format: date-time }
      required: [id, email, name, createdAt]
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        user: { $ref: "#/components/schemas/User" }
      required: [token, user]
    Document:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        pageCount: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, title, pageCount, createdAt]
    Tag:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        documentIds:
          type: array
          items: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, name, documentIds]
    ProfileMaster:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        systemPrompt: { type: string }
      required: [id, name, description]
    Session:
      type: object
      properties:
        id: { type: integer }
        documentId: { type: integer }
        title: { type: string }
        startPage: { type: integer }
        endPage: { type: integer }
        currentPage: { type: integer }
        difficulty: { $ref: "#/components/schemas/Difficulty" }
        profileMasterId: { type: integer, nullable: true }
        language: { $ref: "#/components/schemas/SessionLanguage" }
        createdAt: { type: string, format: date-time }
      required: [id, documentId, title, startPage, endPage, currentPage, difficulty, language, createdAt]
...
    QuizQuestion:
      type: object
      properties:
        id: { type: integer }
        pageNumber: { type: integer }
        question: { type: string }
      required: [id, pageNumber, question]
    QuizConfig:
      type: object
      properties:
        profileMasterId: { type: integer, nullable: true }
        masterName: { type: string, nullable: true }
        difficulty: { $ref: "#/components/schemas/Difficulty" }
        questionCount: { type: integer }
      required: [difficulty, questionCount]
    QuizGenerateResult:
      type: object
      properties:
        questions:
          type: array
          items: { $ref: "#/components/schemas/QuizQuestion" }
        config: { $ref: "#/components/schemas/QuizConfig" }
      required: [questions, config]
    QuizAnswerResult:
      type: object
      properties:
        correct: { type: boolean }
        score: { type: number, minimum: 0, maximum: 1 }
        feedback: { type: string }
      required: [correct, score, feedback]
    QuizAttempt:
      type: object
      properties:
        id: { type: integer }
        pageNumber: { type: integer }
        question: { type: string }
        answer: { type: string }
        correct: { type: boolean }
        score: { type: number }
        feedback: { type: string }
        difficulty: { $ref: "#/components/schemas/Difficulty" }
        masterName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [id, pageNumber, question, answer, correct, score, feedback, difficulty, createdAt]
    QuizReport:
      type: object
      properties:
        attempts:
          type: array
          items: { $ref: "#/components/schemas/QuizAttempt" }
        summary:
          type: object
          properties:
            total: { type: integer }
            correct: { type: integer }
            incorrect: { type: integer }
            averageScore: { type: integer, description: "0-100" }
          required: [total, correct, incorrect, averageScore]
      required: [attempts, summary]
...